Version 1.0
This is the base code for using Python and Splinter to automate a
tedious task. Most functions replace button presses. You must put in the
store numbers you wish to automate in the storeIds array. This isn't
perfected yet, but it works fairly well. This does not work on Upload
Products.